+++ src/main/java/com/github/danielfabela/partpicker/components/Components.java
@@ -1,34 +1,85 @@
 package com.github.danielfabela.partpicker.components;
-import com.github.danielfabela.partpicker.io.IO;
-import com.github.danielfabela.partpicker.menus.Menu;
-import java.util.List;
-import java.io.File;
 
-public abstract class Components {
-    protected String id;
+public abstract class Components implements TaxComputation {
+    protected int id;
+    protected String Part;
     protected String Model;
-    protected String Price;
-    protected double subtotal;
-    protected double taxes;
+    protected String sPrice;
+    protected double dPrice;
+    protected double taxes = 0.0625;
     protected double total;
 
-    public Components(String id, String Model, String Price){
-        this.id = id;
+    public Components(String Part, String Model, String sPrice){
+        this.Part = Part;
         this.Model = Model;
-        this.Price = Price;
+        this.sPrice = sPrice;
+        parseToDouble();
     }
 
-    public Components(String id, String Model, String Price, double subtotal, double taxes, double total){
+    public Components(int id, String Part, String Model, String sPrice, double taxes, double total){
         this.id = id;
+        this.Part = Part;
         this.Model = Model;
-        this.Price = Price;
-        this.subtotal = subtotal;
+        this.sPrice = sPrice;
         this.taxes = taxes;
         this.total = total;
+        parseToDouble();
     }
 
-    //public Components getComponents(String id, String Model, double Price){
-        
-   // }
+    public int getId() {
+		return id;
+    }
     
+    public String getPart() {
+		return Part;
+	}
+
+	public String getModel() {
+		return Model;
+	}
+
+	public String getPrice() {
+		return sPrice;
+	}
+
+	public double getTaxes() {
+		return taxes;
+    }
+    
+    public double getTotal() {
+		return total;
+    }
+    
+    public void setTaxes(double taxes) {
+		this.taxes = taxes;
+    }
+
+	public void setTotal(double total) {
+		this.total = total;
+    }
+    
+    public void parseToDouble(){
+        try{
+            this.dPrice = new Double(sPrice);
+        } catch(NumberFormatException e){
+            System.err.println("Failure to parse string: " + this.sPrice);
+        }
+    }
+
+	@Override
+	public String toString() {
+		StringBuilder out = new StringBuilder();
+		out.append(this.id);
+    out.append(",");
+    out.append(this.Part);
+    out.append(",");
+		out.append(this.Model);
+		out.append(",");
+		out.append(this.sPrice);
+		out.append(",");
+    out.append(this.taxes);
+    out.append(",");
+		out.append(this.total);
+		return out.toString();
+	}
 }