+++ src/main/java/com/github/danielfabela/partpicker/components/TaxComputation.java
@@ -1,0 +1,5 @@
+package com.github.danielfabela.partpicker.components;
+
+public interface TaxComputation {
+    double TotalPrice();
+}