+++ src/main/java/com/github/danielfabela/partpicker/io/Dao.java
@@ -1,0 +1,9 @@
+package com.github.danielfabela.partpicker.io;
+
+import java.util.List;
+
+public interface Dao<E> {
+    void writeAll(List<E> e);
+
+    List<E> File_Read();
+}