+++ src/main/java/com/github/danielfabela/partpicker/io/IO.java
@@ -10,26 +10,27 @@ import java.util.ArrayList;
 import java.util.List;
 
 import com.github.danielfabela.partpicker.components.Components;
+import com.github.danielfabela.partpicker.components.ComponentsFactory;
 
-public class IO{
+public class IO implements Dao<Components>{
     static final File input = new File("PartsInventory.csv");
 
     public IO(){}
 
-    public List<String> File_Read(){
-        List<String> parts = new ArrayList<>();
+    @Override
+    public List<Components> File_Read(){
+        List<Components> parts = new ArrayList<>();
         try (BufferedReader br = new BufferedReader(new FileReader(input))){
+            ComponentsFactory factory = ComponentsFactory.getInstance();
             String line = br.readLine();
             while (line!= null) {
                 // use comma as separator to separate the strings
                 String[] partDetails = line.split(",");
-                String id = partDetails[0];
+                String Part = partDetails[0];
                 String Model = partDetails[1];
-                String Price = partDetails[2];
+                String sPrice = partDetails[2];
                 // add each string into the array list
-                parts.add(id);
-                parts.add(Model);
-                parts.add(Price);
+                parts.add(factory.getComponents(Part, Model, sPrice));
                 // keep reading the next line until the end of file
                 line = br.readLine();
             }
@@ -40,9 +41,10 @@ public class IO{
         } 
         return parts;
     }
-    // Method to write the array list
-    public void writeAll(List<String> parts){
 
+    @Override
+    // Method to write the array list
+    public void writeAll(List<Components> parts){
     }
     // method to take input from the user
     public int MenuInput(){