+++ src/main/java/com/github/danielfabela/partpicker/menus/Menu.java
@@ -7,20 +7,21 @@ public class Menu{
     // Method to display the main Menu
     public void MainMenu(){
         System.out.println("*************************************************************");
-        System.out.println("Welcome to PC Center");
+        System.out.println("Welcome to PC Budget");
         System.out.println("*************************************************************");
         System.out.println("Main Menu");
         System.out.println("0. Quit");
-        System.out.println("1. See Components");
-        System.out.println("2. Choose Components");
-        System.out.println("3. Total Price");
+        System.out.println("1. See Listed Components");
+        System.out.println("2. Calculate Total Price of Each Part");
+        System.out.println("3. Display Total Price for PC");
         System.out.println("*************************************************************");
         System.out.println("Input your choice. Only numbers from 0 to 3.");
     }
 
     public void ComponentsMenu(){
         System.out.println("*************************************************************");
-            System.out.println("Stock Inventory");
+            System.out.println("Items in your list");
+            System.out.println("    Part  |                Model              |  Price  ");
            
     }